Atividade 02

+++ atividades/02/1041_2v.py
@@ -1,5 +1,6 @@
-x = int(input("X: "))
-y = int(input("Y: "))
+x = float(input("X: "))
+y = float(input("Y: "))
+
 
 if (x == 0) or (y == 0):
     if (x != 0):
